requirements/base.txt
# SHA1:b2b2f1d1962ed676253f5dd1723fbf7dbd032985
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
asgiref==3.5.2
    # via django
django==4.1.3
    # via
    #   django-filter
    #   djangorestframework
django-filter==22.1
    # via -r requirements/base.in
djangorestframework==3.14.0
    # via -r requirements/base.in
markdown==3.4.1
    # via -r requirements/base.in
mysqlclient==2.1.1
    # via -r requirements/base.in
pillow==9.3.0
    # via -r requirements/base.in
pygments==2.13.0
    # via -r requirements/base.in
python-dateutil==2.8.2
    # via -r requirements/base.in
pytz==2022.6
    # via djangorestframework
pyyaml==6.0
    # via -r requirements/base.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.3
    # via django
uritemplate==4.1.1
    # via -r requirements/base.in
